+++ shared/src/commonMain/kotlin/eu/heha/conifer/model/database/AppDatabase.kt
@@ -1,11 +1,11 @@
 package eu.heha.conifer.model.database
 
-import androidx.room.AutoMigration
-import androidx.room.ConstructedBy
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import androidx.room.RoomDatabaseConstructor
-import androidx.room.TypeConverters
+import androidx.room3.AutoMigration
+import androidx.room3.ColumnTypeConverters
+import androidx.room3.ConstructedBy
+import androidx.room3.Database
+import androidx.room3.RoomDatabase
+import androidx.room3.RoomDatabaseConstructor
 
 const val DATABASE_NAME = "conifer_database.db"
 
@@ -16,14 +16,14 @@ const val DATABASE_NAME = "conifer_database.db"
         AutoMigration(from = 1, to = 2, spec = Migration1to2::class)
     ]
 )
-@TypeConverters(DatabaseConverters::class)
+@ColumnTypeConverters(DatabaseConverters::class)
 @ConstructedBy(AppDatabaseConstructor::class)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun bitDao(): BitDao
 }
 
 // The Room compiler generates the `actual` implementations.
-@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+@Suppress("NO_ACTUAL_FOR_EXPECT", "EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 expect object AppDatabaseConstructor : RoomDatabaseConstructor<AppDatabase> {
     override fun initialize(): AppDatabase
 }