+++ shared/src/commonMain/kotlin/eu/heha/conifer/model/database/BitDao.kt
@@ -1,8 +1,9 @@
 package eu.heha.conifer.model.database
 
-import androidx.room.Dao
-import androidx.room.Query
-import androidx.room.Upsert
+import androidx.room3.Dao
+import androidx.room3.Delete
+import androidx.room3.Query
+import androidx.room3.Upsert
 import kotlinx.coroutines.flow.Flow
 import kotlin.time.Instant
 
@@ -17,6 +18,9 @@ interface BitDao {
 
     @Upsert
     suspend fun upsert(bit: Bit)
+
+    @Delete
+    suspend fun delete(bit: Bit)
 
     @Query("SELECT * FROM bits WHERE id IN (:bitIds) ORDER BY date DESC, created_at DESC")
     suspend fun getBitsByIds(bitIds: List<String>): List<Bit>