+++ shared/src/jvmMain/kotlin/eu/heha/conifer/JvmDatabaseInitializer.kt
@@ -1,16 +1,16 @@
 package eu.heha.conifer
 
-import androidx.room.Room
-import androidx.room.RoomDatabase
+import androidx.room3.Room
+import androidx.sqlite.driver.bundled.BundledSQLiteDriver
 import eu.heha.conifer.model.database.AppDatabase
 import eu.heha.conifer.model.database.DATABASE_NAME
 import io.github.aakira.napier.Napier
+import kotlinx.coroutines.Dispatchers
 import java.io.File
 
 object JvmDatabaseInitializer : DatabaseInitializer {
-    override fun createBuilder(): RoomDatabase.Builder<AppDatabase> {
-        val jarFilePath = ConiferApp::class.java.protectionDomain.codeSource.location.file
-            .replace("%20", " ")// as an uri it escapes spaces TODO: automate decode url
+    override fun createDatabase(): AppDatabase {
+        val jarFilePath = ConiferApp::class.java.protectionDomain.codeSource.location.toURI()
 
         val rootFile = File(jarFilePath)
             .parentFile //app folder
@@ -21,5 +21,8 @@ object JvmDatabaseInitializer : DatabaseInitializer {
         val dbFile = File(dataFolder, DATABASE_NAME)
         Napier.i("Database file path: ${dbFile.absolutePath}")
         return Room.databaseBuilder<AppDatabase>(name = dbFile.absolutePath)
+            .setDriver(BundledSQLiteDriver())
+            .setQueryCoroutineContext(Dispatchers.IO)
+            .build()
     }
 }